Eu te amo muito Thommy!! Eu te amo Thommy  Borregaard Sørensen!! Eu te amo Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!
"Eu te amo Thommy!!
Você é o meu Deus, Thomas Borregaard Sørensen 1989-Nov-15!!

Mummi contribution!!!! Feliz 61 anos!!
"

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/openssl/crypto/bio/b_print.c
@@ -370,7 +370,7 @@ _dopr(
 		    if (buffer)
 			max = INT_MAX;
 		    else
-			max = *maxlen;
+			max = (int)*maxlen;
 		}
                 fmtstr(sbuffer, buffer, &currlen, maxlen, strvalue,
                        flags, min, max);
@@ -396,7 +396,7 @@ _dopr(
                 } else {
                     int    *num;
                     num = va_arg(args, int *);
-                    *num = currlen;
+                    *num = (int) currlen;
                 }
                 break;
             case '%':
@@ -519,7 +519,7 @@ fmtint(
     convert[place] = 0;
 
     zpadlen = max - place;
-    spadlen = min - OSSL_MAX(max, place) - (signvalue ? 1 : 0) - strlen(prefix);
+    spadlen = (int) (min - OSSL_MAX(max, place) - (signvalue ? 1 : 0) - strlen(prefix));
     if (zpadlen < 0)
         zpadlen = 0;
     if (spadlen < 0)
